Updated "file" class to be able to write arbitray width words

+++ haar3D.h
@@ -13,8 +13,8 @@ struct vali
     double  cz;
 };
 
-void copyAsort(double *VX, double *CX, uint64_t N, vali *C, uint64_t *W, uint64_t *val, uint64_t *ord);
-void haar3D(double *inV, double *inC, uint64_t N, uint64_t depth, double *outCT, double *outW=NULL);
-void inv_haar3D(double *inV, double *inCT, uint64_t N, uint64_t depth, double *outC);
+void copyAsort(double *VX, double *CX, size_t N, vali *C, uint64_t *W, uint64_t *val, uint64_t *ord);
+void haar3D(double *inV, double *inC, size_t N, size_t depth, double *outCT, double *outW=NULL);
+void inv_haar3D(double *inV, double *inCT, size_t N, size_t depth, double *outC);
 
 #endif // EXTRA_H
